refactor: move edit helpers to utils2
