+++ vuu-ui/packages/vuu-utils2/src/edit-utils.ts
@@ -1,0 +1,5 @@
+import type { EditSessionMode } from "@vuu-ui/vuu-data-types";
+
+export const isInlineEditingSession = (
+  mode: EditSessionMode,
+): mode is "inline-all-rows" => mode === "inline-all-rows";
